updated tools

+++ sources/kotlin/io/ByteStreamsKt.java
@@ -22,7 +22,7 @@ public abstract class ByteStreamsKt {
         long j = 0;
         while (i2 >= 0) {
             out.write(bArr, 0, i2);
-            j += i2;
+            j += (long) i2;
             i2 = inputStream.read(bArr);
         }
         return j;
